Also return line1 from definition(::Method, String)
The String variant of `definition` has the ability to back up in the
source file to also extract the signature. As a consequence, it may
be unclear how the first line of the string corresponds to the output
of `whereis`. To resolve the uncertainty, return the first source line
as a second argument from `definition(::Method, String)`.

This is a breaking change.