Don't assume that a method body's file is always the same as the method def
`CodeTracking.whereis` currently assumes that the only thing a LineNumberNode can change
is the line within the function, but that the file is always the same as the file that
the method was defined in (unless there's a macro expansion). That is true for
methods that come from the parser, but not necessarily for those that were generated
programatically, causing `whereis` to return an incorrect file with the correct line
number, causing dowstream issues like https://github.com/JuliaDebug/CassetteOverlay.jl/issues/18.