Provide an inverted linetable for all Julia versions (#135)
https://github.com/JuliaLang/julia/pull/52415 introduced a fundamental
change in how line information is encoded. Because this package delves
into internals, there were a couple tests that failed on Julia 1.12+.
This adds a utility, `CodeTracking.linetable_scopes`, that makes it
easier to work with the new representation across Julia versions.

Importantly, this is "forward-looking," meaning it mimics Julia 1.12+
representations for older Julia versions. That should help with adoption
of the new representation.